Add BlockAuthorCreation method for Posts, and GetAuthorCollection
Adds a method to return 409 Conflict in the case of a post request to an existing author Guid. This overrides the .NET default of returning 404 NotFound, with the more appropriate status code 409. Adds a GetAuthorCollection method, with an ArrayModelBinder.cs helper which handles the binding of the URI id strings to an array of Guids, which become the method's ids parameter.