Add pagination metadata generation, filter authors by genre
Adds a UrlHelper service and injects it into the AuthorsController for generating previous- and next-page links. Adds logic to the controller's GetAuthors method so that pagination query strings passed in via the URI get serialized and returned as metadata. Also adds a property 'Genre' to the AuthorsResourceParameters helper class so that the repository's GetAuthors method can filter for authors based on genre (if genre is specified in the query string).